backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==10.0.0
